fixUnionTestCase
